[fix]: [DEL-3859]: move HttpTask and GitFetchFilesTask to 930-delegat… (#32414)
* [fix]: [DEL-3859]: move GitFetchFilesTask to 930-delegate-tasks

* [fix]: [DEL-3859]: move GitFetchFilesTask to 930-delegate-tasks

* [fix]: [DEL-3859]: move GitFetchFilesTask to 930-delegate-tasks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks

* [fix]: [DEL-3859]: fix pr checks
