+++ 270-delegate-service-app/module-dependency.hash
@@ -1,1 +1,1 @@
-b18a4e041548926527e9ff01db5766700c73e3d2ea7e9cdf7ed6b1ff4564b42b
+54acf64ea365fb7faf0faf5b16d80dcaf0898cbee5fecf977e2a68845e9d48d1
