+++ 400-rest/src/main/java/io/harness/serializer/kryo/ManagerKryoRegistrar.java
@@ -181,7 +181,6 @@ import software.wings.beans.AzureResourceGroup;
 import software.wings.beans.AzureVMSSInfrastructureMapping;
 import software.wings.beans.AzureWebAppInfrastructureMapping;
 import software.wings.beans.Base;
-import software.wings.beans.BastionConnectionAttributes;
 import software.wings.beans.BugsnagConfig;
 import software.wings.beans.CanaryWorkflowExecutionAdvisor;
 import software.wings.beans.CodeDeployInfrastructureMapping;
@@ -199,8 +198,6 @@ import software.wings.beans.EntityVersion;
 import software.wings.beans.Event;
 import software.wings.beans.GcpKubernetesCluster;
 import software.wings.beans.GcpKubernetesInfrastructureMapping;
-import software.wings.beans.GitFetchFilesConfig;
-import software.wings.beans.GitFetchFilesTaskParams;
 import software.wings.beans.GitValidationParameters;
 import software.wings.beans.Graph;
 import software.wings.beans.GraphGroup;
@@ -214,7 +211,6 @@ import software.wings.beans.HostValidationResponse;
 import software.wings.beans.HostValidationTaskParameters;
 import software.wings.beans.InfraMappingSweepingOutput;
 import software.wings.beans.InstanceUnitType;
-import software.wings.beans.KubernetesClusterConfig;
 import software.wings.beans.NameValuePair;
 import software.wings.beans.PcfInfrastructureMapping;
 import software.wings.beans.Permission;
@@ -224,11 +220,9 @@ import software.wings.beans.PhysicalInfrastructureMapping;
 import software.wings.beans.PhysicalInfrastructureMappingWinRm;
 import software.wings.beans.PipelineStageExecutionAdvisor;
 import software.wings.beans.PrometheusConfig;
-import software.wings.beans.RancherConfig;
 import software.wings.beans.RancherKubernetesInfrastructureMapping;
 import software.wings.beans.Role;
 import software.wings.beans.RoleType;
-import software.wings.beans.SSHExecutionCredential;
 import software.wings.beans.ScalyrConfig;
 import software.wings.beans.ServiceInstance;
 import software.wings.beans.ServiceVariable;
@@ -248,7 +242,6 @@ import software.wings.beans.UserInvite;
 import software.wings.beans.UserInviteSource;
 import software.wings.beans.Variable;
 import software.wings.beans.WinRmConnectionAttributes;
-import software.wings.beans.appmanifest.AppManifestKind;
 import software.wings.beans.appmanifest.ApplicationManifest;
 import software.wings.beans.appmanifest.ApplicationManifest.ManifestCollectionStatus;
 import software.wings.beans.appmanifest.ManifestFile;
@@ -370,7 +363,6 @@ import software.wings.beans.template.dto.HarnessImportedTemplateDetails;
 import software.wings.beans.trigger.TriggerCommand.TriggerCommandType;
 import software.wings.beans.utm.UtmInfo;
 import software.wings.beans.yaml.ChangeContext;
-import software.wings.beans.yaml.GitFetchFilesFromMultipleRepoResult;
 import software.wings.beans.yaml.GitPullResult;
 import software.wings.beans.yaml.YamlType;
 import software.wings.cloudprovider.CodeDeployDeploymentInfo;
@@ -387,8 +379,6 @@ import software.wings.delegatetasks.manifest.ManifestCollectionResponse;
 import software.wings.delegatetasks.rancher.RancherResolveClustersResponse;
 import software.wings.delegatetasks.rancher.RancherResolveClustersTaskParameters;
 import software.wings.delegatetasks.validation.capabilities.BasicValidationInfo;
-import software.wings.delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability;
-import software.wings.delegatetasks.validation.capabilities.GitConnectionCapability;
 import software.wings.delegatetasks.validation.capabilities.HelmCommandCapability;
 import software.wings.delegatetasks.validation.capabilities.SSHHostValidationCapability;
 import software.wings.delegatetasks.validation.capabilities.ShellConnectionCapability;
@@ -500,7 +490,6 @@ import software.wings.security.authentication.TwoFactorAuthenticationMechanism;
 import software.wings.security.encryption.secretsmanagerconfigs.CustomSecretsManagerConfig;
 import software.wings.security.encryption.secretsmanagerconfigs.CustomSecretsManagerShellScript;
 import software.wings.security.encryption.secretsmanagerconfigs.CustomSecretsManagerShellScript.ScriptType;
-import software.wings.service.impl.ContainerServiceParams;
 import software.wings.service.impl.MasterUrlFetchTaskParameter;
 import software.wings.service.impl.PerpetualTaskCapabilityCheckResponse;
 import software.wings.service.impl.SlackMessageSenderImpl;
@@ -852,7 +841,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(ManifestFile.class, 5539);
     kryo.register(AzureKubernetesCluster.class, 5538);
     kryo.register(Base.class, 5001);
-    kryo.register(BastionConnectionAttributes.class, 5073);
     kryo.register(BugsnagConfig.class, 5490);
     kryo.register(CanaryWorkflowExecutionAdvisor.class, 5024);
     kryo.register(AbstractCommandUnit.class, 5030);
@@ -919,8 +907,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(ElasticLoadBalancerConfig.class, 5020);
     kryo.register(ElementExecutionSummary.class, 5027);
     kryo.register(GcpKubernetesCluster.class, 5537);
-    kryo.register(GitFetchFilesConfig.class, 5616);
-    kryo.register(GitFetchFilesTaskParams.class, 5575);
     kryo.register(Graph.class, 5060);
     kryo.register(GraphGroup.class, 5063);
     kryo.register(GraphLink.class, 5062);
@@ -933,7 +919,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(KubernetesContainerInfo.class, 5165);
     kryo.register(InstanceUnitType.class, 5221);
     kryo.register(JiraTaskParameters.class, 5579);
-    kryo.register(KubernetesClusterConfig.class, 5244);
     kryo.register(NameValuePair.class, 5226);
     kryo.register(Permission.class, 5310);
     kryo.register(PrometheusConfig.class, 5314);
@@ -946,7 +931,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(ServiceVariable.OverrideType.class, 5361);
     kryo.register(ServiceVariable.Type.class, 5362);
     kryo.register(ServiceVariable.class, 5359);
-    kryo.register(SSHExecutionCredential.class, 5022);
     kryo.register(LdapConnectionSettings.class, 5496);
     kryo.register(LdapGroupResponse.class, 5501);
     kryo.register(LdapGroupSettings.class, 5498);
@@ -965,7 +949,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(Variable.class, 5378);
     kryo.register(WinRmConnectionAttributes.class, 5255);
     kryo.register(ChangeContext.class, 5199);
-    kryo.register(GitFetchFilesFromMultipleRepoResult.class, 5615);
     kryo.register(GitPullResult.class, 5209);
     kryo.register(YamlType.class, 5211);
     kryo.register(CodeDeployDeploymentInfo.class, 5077);
@@ -1105,7 +1088,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(CloudWatchDataCollectionInfo.class, 5317);
     kryo.register(CloudWatchMetric.class, 5318);
     kryo.register(CloudWatchSetupTestNodeData.class, 5494);
-    kryo.register(ContainerServiceParams.class, 5156);
 
     kryo.register(NewRelicMarkerExecutionData.class, 5243);
     kryo.register(PrometheusDataCollectionInfo.class, 5311);
@@ -1265,7 +1247,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(AssignmentTaskResponse.class, 7238);
     kryo.register(SpotInstSetupStateExecutionData.class, 7241);
     kryo.register(SpotinstDeployExecutionSummary.class, 7242);
-    kryo.register(AppManifestKind.class, 7243);
     kryo.register(SpotinstAllPhaseRollbackData.class, 7245);
     kryo.register(NewRelicDataCollectionInfoV2.class, 7247);
     kryo.register(StackDriverMetricDefinition.class, 7249);
@@ -1343,7 +1324,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(ServiceArtifactElements.class, 7342);
     kryo.register(ServiceArtifactVariableElements.class, 7343);
     kryo.register(SpotinstTrafficShiftAlbSwapRoutesExecutionData.class, 7344);
-    kryo.register(ClusterMasterUrlValidationCapability.class, 7345);
     kryo.register(HelmDeployContextElement.class, 7346);
     kryo.register(StackdriverLogGcpConfigTaskParams.class, 7360);
     kryo.register(StackdriverGcpConfigTaskParams.class, 7361);
@@ -1360,7 +1340,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(AwsLambdaDetailsMetricsResponse.class, 7388);
     kryo.register(ServiceNowApprovalParams.class, 7389);
     kryo.register(ShellConnectionCapability.class, 7390);
-    kryo.register(GitConnectionCapability.class, 7391);
     kryo.register(GitValidationParameters.class, 7392);
     kryo.register(AppPermissionSummaryForUI.class, 7395);
     kryo.register(UserRestrictionInfo.class, 7396);
@@ -1579,7 +1558,6 @@ public class ManagerKryoRegistrar implements KryoRegistrar {
     kryo.register(RancherKubernetesInfrastructureMapping.class, 50003);
     kryo.register(RancherResolveClustersTaskParameters.class, 50004);
     kryo.register(RancherResolveClustersResponse.class, 50005);
-    kryo.register(RancherConfig.class, 50006);
     kryo.register(RancherKubernetesInfrastructure.class, 50007);
     kryo.register(RancherKubernetesInfrastructure.ClusterSelectionCriteriaEntry.class, 50008);
     kryo.register(RancherStateExecutionData.class, 50009);
