+++ 400-rest/src/main/java/io/harness/serializer/morphia/ManagerMorphiaRegistrar.java
@@ -176,7 +176,6 @@ import software.wings.beans.BarrierInstance;
 import software.wings.beans.Base;
 import software.wings.beans.BaseFile;
 import software.wings.beans.BasicOrchestrationWorkflow;
-import software.wings.beans.BastionConnectionAttributes;
 import software.wings.beans.BlueGreenOrchestrationWorkflow;
 import software.wings.beans.BugsnagConfig;
 import software.wings.beans.BuildWorkflow;
@@ -215,7 +214,6 @@ import software.wings.beans.Idempotent;
 import software.wings.beans.InformationNotification;
 import software.wings.beans.InfrastructureMapping;
 import software.wings.beans.InfrastructureProvisioner;
-import software.wings.beans.KubernetesClusterConfig;
 import software.wings.beans.LambdaSpecification;
 import software.wings.beans.License;
 import software.wings.beans.ManagerConfiguration;
@@ -242,7 +240,6 @@ import software.wings.beans.ResourceLookup;
 import software.wings.beans.Role;
 import software.wings.beans.RollingOrchestrationWorkflow;
 import software.wings.beans.RuntimeInputsConfig;
-import software.wings.beans.SSHExecutionCredential;
 import software.wings.beans.ScalyrConfig;
 import software.wings.beans.Schema;
 import software.wings.beans.Service;
@@ -422,7 +419,6 @@ import software.wings.beans.trigger.Trigger;
 import software.wings.beans.trigger.TriggerExecution;
 import software.wings.beans.trigger.WebHookTriggerCondition;
 import software.wings.beans.trigger.WorkflowAction;
-import software.wings.beans.yaml.GitFetchFilesFromMultipleRepoResult;
 import software.wings.beans.yaml.GitSuccessFulChangeDetail;
 import software.wings.beans.yaml.HarnessSuccessFulChangeDetail;
 import software.wings.beans.yaml.YamlSuccessfulChange;
@@ -431,8 +427,6 @@ import software.wings.collect.CollectEvent;
 import software.wings.delegatetasks.buildsource.BuildSourceCallback;
 import software.wings.delegatetasks.buildsource.BuildSourceCleanupCallback;
 import software.wings.delegatetasks.buildsource.BuildSourceExecutionResponse;
-import software.wings.delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability;
-import software.wings.delegatetasks.validation.capabilities.GitConnectionCapability;
 import software.wings.delegatetasks.validation.capabilities.HelmCommandCapability;
 import software.wings.delegatetasks.validation.capabilities.SSHHostValidationCapability;
 import software.wings.delegatetasks.validation.capabilities.ShellConnectionCapability;
@@ -618,7 +612,6 @@ import software.wings.sm.states.GcpClusterSetup;
 import software.wings.sm.states.HelmDeployState;
 import software.wings.sm.states.HelmRollbackState;
 import software.wings.sm.states.HttpState;
-import software.wings.sm.states.HttpState.HttpStateExecutionResponse;
 import software.wings.sm.states.InstanaState;
 import software.wings.sm.states.JenkinsState;
 import software.wings.sm.states.KubernetesDeploy;
@@ -1162,7 +1155,6 @@ public class ManagerMorphiaRegistrar implements MorphiaRegistrar {
     w.put("beans.alert.UsageLimitExceededAlert", UsageLimitExceededAlert.class);
     w.put("beans.APMVerificationConfig", APMVerificationConfig.class);
     w.put("beans.BasicOrchestrationWorkflow", BasicOrchestrationWorkflow.class);
-    w.put("beans.BastionConnectionAttributes", BastionConnectionAttributes.class);
     w.put("beans.BlueGreenOrchestrationWorkflow", BlueGreenOrchestrationWorkflow.class);
     w.put("beans.BugsnagConfig", BugsnagConfig.class);
     w.put("beans.BuildWorkflow", BuildWorkflow.class);
@@ -1227,7 +1219,6 @@ public class ManagerMorphiaRegistrar implements MorphiaRegistrar {
     w.put("beans.infrastructure.instance.info.SpotinstAmiInstanceInfo", SpotinstAmiInstanceInfo.class);
     w.put("beans.infrastructure.instance.info.AzureVMSSInstanceInfo", AzureVMSSInstanceInfo.class);
     w.put("beans.infrastructure.instance.info.AzureWebAppInstanceInfo", AzureWebAppInstanceInfo.class);
-    w.put("beans.KubernetesClusterConfig", KubernetesClusterConfig.class);
     w.put("beans.MultiServiceOrchestrationWorkflow", MultiServiceOrchestrationWorkflow.class);
     w.put("beans.PhysicalDataCenterConfig", PhysicalDataCenterConfig.class);
     w.put("beans.PrometheusConfig", PrometheusConfig.class);
@@ -1238,7 +1229,6 @@ public class ManagerMorphiaRegistrar implements MorphiaRegistrar {
     w.put("beans.settings.helm.HttpHelmRepoConfig", HttpHelmRepoConfig.class);
     w.put("beans.SlackConfig", SlackConfig.class);
     w.put("beans.SpotInstConfig", SpotInstConfig.class);
-    w.put("beans.SSHExecutionCredential", SSHExecutionCredential.class);
     w.put("beans.StringValue", StringValue.class);
     w.put("beans.template.artifactsource.ArtifactSourceTemplate", ArtifactSourceTemplate.class);
     w.put("beans.template.artifactsource.CustomArtifactSourceTemplate", CustomArtifactSourceTemplate.class);
@@ -1253,17 +1243,13 @@ public class ManagerMorphiaRegistrar implements MorphiaRegistrar {
     w.put("beans.trigger.WebHookTriggerCondition", WebHookTriggerCondition.class);
     w.put("beans.trigger.WorkflowAction", WorkflowAction.class);
     w.put("beans.WinRmConnectionAttributes", WinRmConnectionAttributes.class);
-    w.put("beans.yaml.GitFetchFilesFromMultipleRepoResult", GitFetchFilesFromMultipleRepoResult.class);
     w.put("beans.yaml.GitSuccessFulChangeDetail", GitSuccessFulChangeDetail.class);
     w.put("beans.yaml.HarnessSuccessFulChangeDetail", HarnessSuccessFulChangeDetail.class);
     w.put("collect.ArtifactCollectionCallback", ArtifactCollectionCallback.class);
     w.put("delegatetasks.buildsource.BuildSourceCallback", BuildSourceCallback.class);
     w.put("delegatetasks.buildsource.BuildSourceCleanupCallback", BuildSourceCleanupCallback.class);
     w.put("delegatetasks.buildsource.BuildSourceExecutionResponse", BuildSourceExecutionResponse.class);
-    w.put("delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability",
-        ClusterMasterUrlValidationCapability.class);
     w.put("delegatetasks.validation.capabilities.HelmCommandCapability", HelmCommandCapability.class);
-    w.put("delegatetasks.validation.capabilities.GitConnectionCapability", GitConnectionCapability.class);
     w.put("delegatetasks.validation.capabilities.SSHHostValidationCapability", SSHHostValidationCapability.class);
     w.put("delegatetasks.validation.capabilities.WinrmHostValidationCapability", WinrmHostValidationCapability.class);
     w.put("delegatetasks.validation.capabilities.ShellConnectionCapability", ShellConnectionCapability.class);
@@ -1407,7 +1393,6 @@ public class ManagerMorphiaRegistrar implements MorphiaRegistrar {
     w.put("sm.states.HelmDeployState", HelmDeployState.class);
     w.put("sm.states.HelmRollbackState", HelmRollbackState.class);
     w.put("sm.states.HttpState", HttpState.class);
-    w.put("sm.states.HttpState$HttpStateExecutionResponse", HttpStateExecutionResponse.class);
     w.put("sm.states.InstanaState", InstanaState.class);
     w.put("sm.states.JenkinsState", JenkinsState.class);
     w.put("sm.states.k8s.K8sApplyState", K8sApplyState.class);
