+++ 930-delegate-tasks/src/main/java/io/harness/mappers/AwsSMConfigMapper.java
