+++ 930-delegate-tasks/src/main/java/io/harness/serializer/morphia/DelegateTasksMorphiaRegistrar.java
@@ -12,33 +12,48 @@ import io.harness.morphia.MorphiaRegistrar;
 import io.harness.morphia.MorphiaRegistrarHelperPut;
 
 import software.wings.beans.AppDynamicsConfig;
+import software.wings.beans.AwsSecretsManagerConfig;
 import software.wings.beans.AzureConfig;
+import software.wings.beans.AzureVaultConfig;
 import software.wings.beans.BambooConfig;
 import software.wings.beans.BaseVaultConfig;
+import software.wings.beans.BastionConnectionAttributes;
+import software.wings.beans.CyberArkConfig;
 import software.wings.beans.DockerConfig;
 import software.wings.beans.DynaTraceConfig;
 import software.wings.beans.EcrConfig;
 import software.wings.beans.ElkConfig;
 import software.wings.beans.GcpConfig;
+import software.wings.beans.GcpKmsConfig;
+import software.wings.beans.GcpSecretsManagerConfig;
 import software.wings.beans.HostConnectionAttributes;
 import software.wings.beans.InstanaConfig;
 import software.wings.beans.JiraConfig;
+import software.wings.beans.KmsConfig;
+import software.wings.beans.KubernetesClusterConfig;
+import software.wings.beans.LocalEncryptionConfig;
 import software.wings.beans.NewRelicConfig;
 import software.wings.beans.PcfConfig;
+import software.wings.beans.SSHExecutionCredential;
 import software.wings.beans.SSHVaultConfig;
+import software.wings.beans.SecretManagerRuntimeParameters;
 import software.wings.beans.ServiceNowConfig;
 import software.wings.beans.SftpConfig;
 import software.wings.beans.SmbConfig;
 import software.wings.beans.SplunkConfig;
 import software.wings.beans.SumoConfig;
+import software.wings.beans.VaultConfig;
 import software.wings.beans.appmanifest.HelmChart;
 import software.wings.beans.artifact.ArtifactFile;
 import software.wings.beans.config.ArtifactoryConfig;
 import software.wings.beans.config.LogzConfig;
 import software.wings.beans.config.NexusConfig;
 import software.wings.beans.settings.azureartifacts.AzureArtifactsPATConfig;
+import software.wings.beans.yaml.GitFetchFilesFromMultipleRepoResult;
 import software.wings.delegatetasks.cv.beans.CustomLogResponseMapper;
 import software.wings.delegatetasks.validation.DelegateConnectionResult;
+import software.wings.delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability;
+import software.wings.delegatetasks.validation.capabilities.GitConnectionCapability;
 import software.wings.helpers.ext.mail.SmtpConfig;
 import software.wings.service.impl.analysis.CustomLogDataCollectionInfo;
 import software.wings.service.impl.analysis.DataCollectionTaskResult;
@@ -54,6 +69,15 @@ public class DelegateTasksMorphiaRegistrar implements MorphiaRegistrar {
   public void registerClasses(Set<Class> set) {
     set.add(NewRelicMetricDataRecord.class);
     set.add(DelegateConnectionResult.class);
+    set.add(AwsSecretsManagerConfig.class);
+    set.add(AzureVaultConfig.class);
+    set.add(CyberArkConfig.class);
+    set.add(GcpKmsConfig.class);
+    set.add(GcpSecretsManagerConfig.class);
+    set.add(KmsConfig.class);
+    set.add(LocalEncryptionConfig.class);
+    set.add(VaultConfig.class);
+    set.add(SecretManagerRuntimeParameters.class);
     set.add(HelmChart.class);
     set.add(BaseVaultConfig.class);
     set.add(SecretManagerConfig.class);
@@ -75,7 +99,23 @@ public class DelegateTasksMorphiaRegistrar implements MorphiaRegistrar {
     w.put("service.impl.elk.ElkDataCollectionInfo", ElkDataCollectionInfo.class);
     w.put("service.impl.logz.LogzDataCollectionInfo", LogzDataCollectionInfo.class);
     w.put("beans.EcrConfig", EcrConfig.class);
+    w.put("beans.AwsSecretsManagerConfig", AwsSecretsManagerConfig.class);
+    w.put("beans.AzureVaultConfig", AzureVaultConfig.class);
+    w.put("beans.CyberArkConfig", CyberArkConfig.class);
+    w.put("beans.GcpKmsConfig", GcpKmsConfig.class);
+    w.put("beans.GcpSecretsManagerConfig", GcpSecretsManagerConfig.class);
+    w.put("beans.KmsConfig", KmsConfig.class);
+    w.put("beans.LocalEncryptionConfig", LocalEncryptionConfig.class);
     w.put("beans.SecretManagerConfig", SecretManagerConfig.class);
+    w.put("beans.VaultConfig", VaultConfig.class);
+    w.put("beans.BastionConnectionAttributes", BastionConnectionAttributes.class);
+    w.put("beans.HostConnectionAttributes", HostConnectionAttributes.class);
+    w.put("beans.SSHExecutionCredential", SSHExecutionCredential.class);
+    w.put("delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability",
+        ClusterMasterUrlValidationCapability.class);
+    w.put("beans.yaml.GitFetchFilesFromMultipleRepoResult", GitFetchFilesFromMultipleRepoResult.class);
+    w.put("beans.KubernetesClusterConfig", KubernetesClusterConfig.class);
+    w.put("delegatetasks.validation.capabilities.GitConnectionCapability", GitConnectionCapability.class);
     w.put("beans.SSHVaultConfig", SSHVaultConfig.class);
     w.put("beans.BaseVaultConfig", BaseVaultConfig.class);
     w.put("beans.GcpConfig", GcpConfig.class);
@@ -88,7 +128,6 @@ public class DelegateTasksMorphiaRegistrar implements MorphiaRegistrar {
     w.put("beans.settings.azureartifacts.AzureArtifactsPATConfig", AzureArtifactsPATConfig.class);
     w.put("beans.SftpConfig", SftpConfig.class);
     w.put("beans.InstanaConfig", InstanaConfig.class);
-    w.put("beans.HostConnectionAttributes", HostConnectionAttributes.class);
     w.put("beans.PcfConfig", PcfConfig.class);
     w.put("beans.AzureConfig", AzureConfig.class);
     w.put("beans.BambooConfig", BambooConfig.class);
