+++ 930-delegate-tasks/src/main/java/software/wings/beans/BastionConnectionAttributes.java
