+++ 930-delegate-tasks/src/main/java/software/wings/beans/VaultConfig.java
