+++ 930-delegate-tasks/src/main/java/software/wings/beans/appmanifest/AppManifestKind.java
@@ -11,9 +11,9 @@ import static io.harness.annotations.dev.HarnessModule._950_DELEGATE_TASKS_BEANS
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 import static io.harness.pcf.model.PcfConstants.VARS_YML;
 
-import static software.wings.beans.appmanifest.ManifestFile.VALUES_YAML_KEY;
 import static software.wings.beans.yaml.YamlConstants.KUSTOMIZE_PATCHES_FILE;
 import static software.wings.beans.yaml.YamlConstants.OC_PARAMS_FILE;
+import static software.wings.beans.yaml.YamlConstants.VALUES_YAML_KEY;
 
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.annotations.dev.TargetModule;
