+++ 930-delegate-tasks/src/main/java/software/wings/delegatetasks/delegatecapability/CapabilityHelper.java
@@ -27,6 +27,7 @@ import io.harness.delegate.beans.executioncapability.SocketConnectivityExecution
 import io.harness.delegate.capability.EncryptedDataDetailsCapabilityHelper;
 import io.harness.delegate.task.TaskParameters;
 import io.harness.delegate.task.mixin.HttpConnectionExecutionCapabilityGenerator;
+import io.harness.delegate.task.mixin.ProcessExecutorCapabilityGenerator;
 import io.harness.expression.ExpressionEvaluator;
 import io.harness.security.encryption.EncryptableSettingWithEncryptionDetails;
 import io.harness.security.encryption.EncryptedDataDetail;
@@ -278,5 +279,45 @@ public class CapabilityHelper {
       log.error("Failed to construct uri for {} repo", repoUrl, e);
       return repoUrl;
     }
+  }
+
+  public static List<ExecutionCapability> generateExecutionCapabilitiesForProcessExecutor(String category,
+      List<String> processExecutorArguments, List<EncryptedDataDetail> encryptedDataDetails,
+      ExpressionEvaluator maskingEvaluator) {
+    List<ExecutionCapability> executionCapabilities = new ArrayList<>();
+    executionCapabilities.add(
+        ProcessExecutorCapabilityGenerator.buildProcessExecutorCapability(category, processExecutorArguments));
+
+    if (isNotEmpty(encryptedDataDetails)) {
+      List<ExecutionCapability> capabilitiesForEncryption =
+          EncryptedDataDetailsCapabilityHelper.fetchExecutionCapabilitiesForEncryptedDataDetails(
+              encryptedDataDetails, maskingEvaluator);
+      if (isNotEmpty(capabilitiesForEncryption)) {
+        executionCapabilities.addAll(capabilitiesForEncryption);
+      }
+    }
+    return executionCapabilities;
+  }
+
+  public static List<ExecutionCapability> generateExecutionCapabilitiesForTerraform(
+      List<EncryptedDataDetail> encryptedDataDetails, ExpressionEvaluator maskingEvaluator) {
+    List<String> processExecutorArguments = new ArrayList<>();
+    processExecutorArguments.add("/bin/sh");
+    processExecutorArguments.add("-c");
+    processExecutorArguments.add("terraform --version");
+
+    return generateExecutionCapabilitiesForProcessExecutor(
+        TERRAFORM, processExecutorArguments, encryptedDataDetails, maskingEvaluator);
+  }
+
+  public static List<ExecutionCapability> generateExecutionCapabilitiesForTerragrunt(
+      List<EncryptedDataDetail> encryptedDataDetails, ExpressionEvaluator maskingEvaluator) {
+    List<String> processExecutorArguments = new ArrayList<>();
+    processExecutorArguments.add("/bin/sh");
+    processExecutorArguments.add("-c");
+    processExecutorArguments.add("terragrunt --version");
+
+    return generateExecutionCapabilitiesForProcessExecutor(
+        TERRAGRUNT, processExecutorArguments, encryptedDataDetails, maskingEvaluator);
   }
 }
