+++ 930-delegate-tasks/src/main/java/software/wings/delegatetasks/validation/capabilities/GitConnectionCapability.java
