+++ 930-delegate-tasks/src/main/java/software/wings/service/impl/ContainerServiceParams.java
@@ -28,7 +28,7 @@ import software.wings.beans.GcpConfig;
 import software.wings.beans.KubernetesClusterConfig;
 import software.wings.beans.RancherConfig;
 import software.wings.beans.SettingAttribute;
-import software.wings.delegatetasks.delegatecapability.CapabilityHelper;
+import software.wings.delegatetasks.utils.CapablityUtility;
 import software.wings.delegatetasks.validation.capabilities.ClusterMasterUrlValidationCapability;
 import software.wings.settings.SettingValue;
 
@@ -80,7 +80,7 @@ public class ContainerServiceParams implements ExecutionCapabilityDemander {
 
     List<ExecutionCapability> executionCapabilities = new ArrayList<>();
     if (value instanceof RancherConfig) {
-      return CapabilityHelper.generateDelegateCapabilities(value, encryptionDetails, maskingEvaluator);
+      return CapablityUtility.generateDelegateCapabilities(value, encryptionDetails, maskingEvaluator);
     } else if (value instanceof AwsConfig) {
       return value.fetchRequiredExecutionCapabilities(maskingEvaluator);
     } else if (value instanceof KubernetesClusterConfig
